do some updates

+++ routes/user_routes.py
@@ -3,6 +3,11 @@ from sqlalchemy.orm import Session
 from app.database import SessionLocal
 from models.user_model import User
 from schemas.user_schema import UserCreate
+from utils.security import hash_password
+from fastapi import HTTPException
+from schemas.user_schema import UserLogin
+from utils.security import verify_password
+from auth.jwt_handler import create_access_token
 
 router = APIRouter(prefix="/users", tags=["Users"])
 
@@ -16,17 +21,23 @@ def get_db():
 
 
 # Create User
+
+
 @router.post("/")
 def create_user(user: UserCreate, db: Session = Depends(get_db)):
     new_user = User(
         username=user.username,
         email=user.email,
-        password=user.password
+        password=hash_password(user.password)
     )
     db.add(new_user)
     db.commit()
     db.refresh(new_user)
-    return new_user
+    return {
+        "id": new_user.id,
+        "username": new_user.username,
+        "email": new_user.email
+    }
 
 
 # Get all users
@@ -43,10 +54,45 @@ def get_user(user_id: int, db: Session = Depends(get_db)):
     return user
 
 
+@router.post("/login")
+def login(user: UserLogin, db: Session = Depends(get_db)):
+
+    # 1. get user
+    db_user = db.query(User).filter(User.username == user.username).first()
+
+    if not db_user:
+        raise HTTPException(status_code=404, detail="User not found")
+
+    # 2. verify password
+    if not verify_password(user.password, db_user.password):
+        raise HTTPException(status_code=401, detail="Invalid credentials")
+
+    # 3. create token
+    token = create_access_token({
+        "sub": db_user.username,
+        "user_id": db_user.id
+    })
+
+    # 4. return token
+    return {
+        "access_token": token,
+        "token_type": "bearer"
+    }
+
+
+
+
 # Delete user
+
 @router.delete("/{user_id}")
 def delete_user(user_id: int, db: Session = Depends(get_db)):
     user = db.query(User).filter(User.id == user_id).first()
+
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+
     db.delete(user)
     db.commit()
+
     return {"message": "User deleted"}
+
